isUpdated, new `CommissionCurrent`
